File Manager service, upgrade to Dotnet Core 3.1

+++ cllc-public-app/ViewModels/FileSystemItem.cs
@@ -7,7 +7,7 @@ namespace Gov.Lclb.Cllb.Public.ViewModels
         public string id { get; set; }
         public string name { get; set; }
         public string documenttype { get; set; }
-        public int size { get; set; }
+        public long size { get; set; }
         public string serverrelativeurl { get; set; }
         public DateTime timecreated { get; set; }
         public DateTime timelastmodified { get; set; }
